EICNET-1143: Move Message Creator classes to Service folder and call them as a service.
