+++ lib/modules/eic_messages/eic_messages.module
@@ -23,8 +23,8 @@ use Drupal\node\Entity\NodeType;
  * Implements hook_ENTITY_TYPE_insert().
  */
 function eic_messages_group_insert(EntityInterface $entity) {
-  /** @var Drupal\eic_messages\Hooks\GroupMessageCreator $class */
-  $class = \Drupal::classResolver(GroupMessageCreator::class);
+  /** @var Drupal\eic_messages\Service\GroupMessageCreator $class */
+  $class = \Drupal::service('eic_messages.message_creator.group');
   $class->groupInsert($entity);
 }
 
@@ -32,8 +32,8 @@ function eic_messages_group_insert(EntityInterface $entity) {
  * Implements hook_ENTITY_TYPE_update().
  */
 function eic_messages_group_update(EntityInterface $entity) {
-  /** @var Drupal\eic_messages\Hooks\GroupMessageCreator $class */
-  $class = \Drupal::classResolver(GroupMessageCreator::class);
+  /** @var Drupal\eic_messages\Service\GroupMessageCreator $class */
+  $class = \Drupal::service('eic_messages.message_creator.group');
   $class->groupUpdate($entity);
 }
 
@@ -41,8 +41,8 @@ function eic_messages_group_update(EntityInterface $entity) {
  * Implements hook_ENTITY_TYPE_insert().
  */
 function eic_messages_group_content_insert(EntityInterface $entity) {
-  /** @var Drupal\eic_messages\Hooks\GroupContentMessageCreator $class */
-  $class = \Drupal::classResolver(GroupContentMessageCreator::class);
+  /** @var Drupal\eic_messages\Service\GroupContentMessageCreator $class */
+  $class = \Drupal::service('eic_messages.message_creator.group_content');
   $class->groupContentInsert($entity);
 }
 
@@ -71,8 +71,8 @@ function eic_messages_request_insert(
   ContentEntityInterface $entity,
   string $type
 ) {
-  /** @var Drupal\eic_messages\Hooks\RequestMessageCreator $class */
-  $class = \Drupal::classResolver(RequestMessageCreator::class);
+  /** @var Drupal\eic_messages\Service\RequestMessageCreator $class */
+  $class = \Drupal::service('eic_messages.message_creator.request');
   $class->requestInsert($flag, $entity, $type);
 }
 
@@ -84,8 +84,8 @@ function eic_messages_request_close(
   ContentEntityInterface $entity,
   string $type
 ) {
-  /** @var Drupal\eic_messages\Hooks\RequestMessageCreator $class */
-  $class = \Drupal::classResolver(RequestMessageCreator::class);
+  /** @var Drupal\eic_messages\Service\RequestMessageCreator $class */
+  $class = \Drupal::service('eic_messages.message_creator.request');
   $class->requestClose($flag, $entity, $type);
 }
